Move TableSettings into persistence package.
