+++ tests/integration/logicblocks/event/processing/broker/subscriptions/stores/state/test_postgres.py
@@ -5,7 +5,7 @@ import pytest_asyncio
 from psycopg import AsyncConnection, abc, sql
 from psycopg_pool import AsyncConnectionPool
 
-from logicblocks.event.db import PostgresConnectionSettings
+from logicblocks.event.persistence.postgres import ConnectionSettings
 from logicblocks.event.processing import (
     EventSubscriptionState,
     EventSubscriptionStateChange,
@@ -24,7 +24,7 @@ from logicblocks.event.testsupport import (
 )
 from logicblocks.event.types import CategoryIdentifier
 
-connection_settings = PostgresConnectionSettings(
+connection_settings = ConnectionSettings(
     user="admin",
     password="super-secret",
     host="localhost",
