+++ tests/integration/logicblocks/event/store/adapters/test_postgres.py
@@ -9,7 +9,7 @@ from psycopg import AsyncConnection, abc, sql
 from psycopg.rows import class_row
 from psycopg_pool import AsyncConnectionPool
 
-from logicblocks.event.db import PostgresConnectionSettings
+from logicblocks.event.persistence.postgres import ConnectionSettings
 from logicblocks.event.store.adapters import (
     EventSerialisationGuarantee,
     EventStorageAdapter,
@@ -39,7 +39,7 @@ from logicblocks.event.testsupport import (
 )
 from logicblocks.event.types import StoredEvent, identifier
 
-connection_settings = PostgresConnectionSettings(
+connection_settings = ConnectionSettings(
     user="admin",
     password="super-secret",
     host="localhost",
